Add quotient and remainder primitives to the compiler
This change adds quotient and remainder primitives to the
compiler. These primitives are needed for number-theoretic integer
division.

The difference between the new primitives and the old ones is that
`quot` always rounds towards zero while `div` rounds towards minus
infinity. Moreover, `rem` has the same sign as the dividend, and `mod`
has the same sign as the divisor.