add prim__getNullAnyPtr and prim__getNullString
This will enable us to get null pointers in Idris2 so that we can pass
them on to other c methods. This is useful when creating FFI bindings
for other libraries that expect optional null pointers to be passed into
the arguments.

For example, the samples/FFI-readline code uses
nullString, a similar method to get a null pointer in Idris2.

https://github.com/idris-lang/Idris2/blob/3d53c2874b4dcfe8ca4b6687b384be261faf74d4/samples/FFI-readline/readline_glue/idris_readline.c#L32-L34