add support for more casts from and to BitsN types
closes #545

Adds primitive casts between different `BitsN` types as well as for the
`Int` and `Integer` type.

`Cast` instances from and to `BitsN` types have been added to the `prelude`.

The scheme backends (Chez, Racket, Gambit) contain functions for
"narrowing" the bit width of integers, widening is a no-op.
The ES backend behaves similarly.