[ breaking ] remove List1 related ambiguities
The constructor for List1 was named (::) so that we could write [x,y,z]
for List1 literals. However that means that every [x,y,z] or (x :: xs)
is ambiguous (it could be either a List or a List1) and can only be
disambiguated if there is a type annotation around.

This is really annoying because if we apply an overloaded function to
one such expression (e.g. show [x,y,z]) then we need to manually
disambiguate!

Similarly, `toList` is a Foldable method and having it overloaded with
List1's function to forget a `List1` back into a `List` is really
annoying.