Use collect-rendezvous to trigger a collection
`collect` requires that only one thread is active, and will
throw an exception otherwise. This is hard to ensure outside
collect-request-handler. `collect-rendezvous` will handle that that and
run the request handler.

This way we can't pass a generation number to collect. That could be done
a boolean parameter saying if we're exiting.