[ new ] do not implicitly generalise names in scope
The fact that

```
mapId : (xs : List a) -> map id xs === xs
```

used to mean

```
mapId : {0 id : a -> b} -> (xs : List a) -> map id xs === xs
```

depiste the fact a function `id : a -> a` is in scope was very confusing
for newcomers (and in obscure cases, experienced users too). With this
patch we only generalise over the variables that are not visible in the
current scope.