Added regression test.
I've added a regression test for this patch. Note, that the test is
slightly different from the error generating code attached to the
corresponding issue (#658).
The reason for this is that despite this patch, interfaces are still
somewhat broken in Idris 2 and that the example given in issue #658
still doesn't type check even with this patch.
In the test I've added I've made the variable auto implicit instead of
mere implicit, this prevents the triggering of the other bug still
present.