Implement a rudimentary local hints mechanism
Each type hint also stores the depth in the environment in which it
was defined, and during proof search, we lower this definition back to
the current environment by applying it to the initial segment of the
environment of that depth.

This is still incomplete:

1. The local hints should be removed once we finished elaborating +
type-checking the local definition. This isn't straightforward because
the elaboration/type-checking of some parts in the scope of the local
definition might be delayed until after we finish elaborating the let
declaration itself.

2. The same mechanism should probably be re-implemented in
TTImp.Interactive.ExprSearch