Allow `HasIO` to not depend on `Monad`.
The definition of `HasIO` depends on `Monad` to provide sequencing of `HasIO` like things.
A strict dependency on `Monad` makes it difficult to provide correct implementatons of `HasIO` for indexed computation contexts in which there are type-level operations.
Idris2's standard interface definitions are, unfortunatly, too restrictive and will pin type-level values unnecessarily as such making the wrong assumptions when defining Applicative, Monad, or even Show implementations.
Moreover, the type required by the interface is not the last type-level parameter.

Here we remove the explicit dependency, and provide an alias (`HasMonadIO`) to regain the dependency and thus allow one to regain 'Do-notation' by overloading `(>>=)` and `pure`.