Save types from 'where' and 'let' in the metadata
More generally, save the types of any variable found in the nested
names. In these cases, the lookup would return Just and getNameType,
a function that also added the variable to the metadata, was not called.
Now the Nothing case also adds it. This affects primarily vars defined
in where clauses and in let clauses that had an explicit type.