Fix TypeAt in the type signature of a function
Previously, the file context for the type signature as a whole would be
used for the name of the function, making TypeAt anywhere within the
signature return the type of the function, instead of the type of the
name the cursor was on.

With this commit, a file context just for the function name is properly
tracked and used. This required updating a test for reflection, as the
ITy node can now have an additional file context, passed as `EmptyFC` in
the test.