Add the LHS of `let` inside `do` to the metadata
For example, :typeat in the name `x` in

```
f : a
f = do
  let x = 5
  ?a
```

used to give UndefinedName. Now, the type of the variable is correctly
reported. `let`s that pattern matched on the variable already worked
properly as they use the constructor DoLetPat and it's `PTerm`s already
hold the necessary file contexts.