Don't add implicits after non-given explicits.
The function `TTImp.TTImp.implicitsAs` is changed so that implicit
arguments that come after explicit ones that are not given are no longer
added.

The reasoning behind this is that the former behaviour---of adding all
missing implicits---makes `implicitsAs` return malformed terms in
certain cases where the underlying function type has implicit arguments
that come after explicit ones. For example, the interface declaration

    interface Foo a where
      bar : a -> {auto ok : ()} -> a

and the function declaration

    foo : Void -> {auto ok : ()} -> Void
    foo = ?foo_hole

result in the errors

    ok is not a valid argument in bar

and

    ok is not a valid argument in foo

respectively, because the auto-implicit is added during elaboration. For
instance, in the second example the left hand side of the pattern
matching declaration

    foo = ?foo_hole

gets replaced by `foo {ok=ok}` (in `implicitsAs`), which is an invalid
function call (according to `TTImp.Elab.App.checkAppWith`).

The changes made are as follows.

The functions `TTImp.TTImp.setAs` and `TTImp.TTImp.findImps` now both
take a new argument `es` of type `List (Maybe Name)`.

In `setAs`, this variable is used to record the explicit arguments which
are explicitly given in the left hand side of the pattern matching
declaration.

This variable is then passed onto `findImps`, which uses it to determine
when it encounters an explicit variable not given and stop searching for
implicits.