Fix file context for argument pattern in lambdas
When desugaring a `PLam` (high level syntax lambda), a variable called
"lamc" is created. Previously, it was using the file context of the
argument to the lambda, obstructing the variables bound by the argument
pattern in the metadata. Now, "lamc" has an empty file context
(`EmptyFC`) as should all variables that don't exist in the original
file (those that are generated by the compiler).