Track the FC for the name in a do bind
Now the name to the left of the arrow in a bind (inside a do block) is
properly placed in the metadata.

When lambdas started adding their arguments to the metadata, the lhs
of do binds (that desugar to lambdas) automatically started being added
as well, although with a file context that corresponds to the whole
expression (the left and the right of the `<-` arrow), obstructing the
types of variables at the rhs. This fixes this problem.