Use empty file context when desugaring do bind
It's necessary that compiler generated variables and lambdas that bind
compiler generated variables have empty file contexts, otherwise they
pollute the metadata with names that don't actually exist in the source
file.

An underscore variable is generated by desugaring a do bind that
pattern matches in the left side of the arrow, `S n <- ?x` for example.
It was being created with a non empty file context, covering expressions
on the right side of the arrow as well.