Use the correct FC when checking application
`checkApp` was using the file context for the whole application as the
file context for the variable being applied. This sometimes made :typeat
report the type of a function when asked the type of an argument to such
function.

This was what was causing the test for :typeat with !-syntax to fail.