Rule out `EmptyFC`s from the metadata mappings
For this purpose, a `NonEmptyFC` type was created, a type for file
contexts that cannot be empty.

The projections from `FC` were replaced by projections from
`NonEmptyFC`, as the original ones had to "silently" return values
such as empty string and zero when called on an empty file context
in order to satisfy the type signature.

The functions that relied on the old projections now handle the EmptyFC
case explicitly.