rtassert and panic function
Perform runtime assertion. Two version of the function:

* `rtassert` and `panic` which are `IO ()`
* `unsafeRtassert` and `unsafePanic` to be used in pure expressions

Sometimes it is hard or not possible to proove correctness, and
proper error propagation makes program clumsy.  `unsafeRtassert`
and `unsafePanic` are an escape hatch.

Have not found a way to write a test for this code (seems like libs
is not tested), so I tested manually with:

```
module A

import Debug.Panic

foo : ()
foo = unsafeRtassert False "rrr" ()

main : IO ()
main = do
    putStrLn "aaa"
    let () = foo
    putStrLn "bbb"
```

```
aaa
assertion failed: rrr
```