Track file contexts for App nodes in normal form
In `Term`, an application such as `f x y z` is represented as
nested `App` nodes, like `App (App (App f x) y) z` (Ommiting the other
parameters to App). In a normal form `NF`, however, such application is
stored as a single `NApp` node, with a list of closures, one for each
argument.

Since both representations store one file context per node, the file
contexts for most of the `App` nodes were lost when converting to normal
form.

When converting a normal form back to a term (quoting), the file context
for the head of the application became the file context for all of the
`App` nodes. When the file context for functions was corrected in a
previous commit, this problem caused the error message for unreachable
cases to highlight just the name of the function instead of the entire
case, due to the wrong FC, as the lhs of functions is normalized then
quoted back by `normaliseHoles`. The wrong highlighting made test
coverage009 fail.

With this commit, the `NApp` node stores a `List (FC, Closure vars)`,
pairing each argument with the file context of it's original `App` node,
preserving the file context information. It is then used when quoting
and mostly ignored elsewhere.

coverage009 now passes again.