Run super-linter in all branches
Use the same trigger as in the rest of jobs.

Before submitting a PR, commit is pushed to a private branch where
the jobs are executed. Super-linter should be executed in these
jobs as well to get signal about linting issue before submitting a
PR, to reduce noise to PR reviewers.