Introduce IDE Protocol Versioning.
Before the IDE Protocol allowed clients to query the version of Idris being used.
When there was a single version of Idris this was the same as querying the version of the IDE Protocol being used.
Various clients (Emacs) uses this to keep track of the version of the IDE Protocol it is used.

With Idris2 and now different protocol implementations and extensions such querying is not useful.

We introduce a notion of IDE Protocol version separate from the compiler version number.
This will allow IDE Clients to differentiate between versions of the protocol used by different versions of Idris.

This is a reasonable first step in supporting compatibility between clients, versions of Idris, and versions of the IDE Protocol.

By default Idris1 will have protocol version 0 and Idris2 protocol version 1.