[ fix ] bring `commit` back
The silent expansion of

```idris
do p
   q
```

to

```
p >>= \ _ => q
```

acted like a pattern-synonym when building these grammars and thus
did not get in the way of termination / productivity checking. When
we introduce (>>), it breaks this convenient feature and to recover
it we need to introduce new constructors.

It's unfortunate but not too much trouble.