Make idris2_createDir default to mode 0755 on unix-like platforms
Without this change, Idris creates directories without an executable bit
set for group and others. This prevents others and group from accessing
entries within in this directory.

I noticed this while working on an Idris 2 package for the Alpine Linux
distribution. The Idris 2 --install command uses this function for
creating directories containing the Idris 2 modules. If these
directories are owned by root:root, non-root users cannot load Idris
modules. For this reason, the Idris 2 REPL also fails to start with the
following error:

	Welcome to Idris 2.  Enjoy yourself!
	Uncaught error: (implicit):1:1--1:1:Prelude not found

While this can be fixed by changing the directory permissions from the
package recipe, I believe 0755 to be the more sensible default.